+++ src/main/java/racingcar/domains/error/ErrorMessageGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domains.error;
 
 public class ErrorMessageGenerator {
     private static final String format = "[ERROR] %s";
