+++ src/main/java/racingcar/domains/validate/AttemptNumberInputValidator.java
@@ -1,6 +1,6 @@
 package racingcar.domains.validate;
 
-import racingcar.ErrorMessageGenerator;
+import racingcar.domains.error.ErrorMessageGenerator;
 import racingcar.domains.error.RacingCarGameException;
 
 public class AttemptNumberInputValidator implements Validator<String> {
