+++ src/main/java/racingcar/domains/validate/CarNamesInputValidator.java
@@ -1,10 +1,9 @@
 package racingcar.domains.validate;
 
-import racingcar.ErrorMessageGenerator;
+import racingcar.domains.error.ErrorMessageGenerator;
 import racingcar.domains.error.RacingCarGameException;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.regex.Pattern;
 
 public class CarNamesInputValidator implements Validator<String> {
